Extract pointCloudEditorProvider.ts clusters into src/providerHandlers/
Applies the same cohesive-cluster extraction pattern used on main.ts:
binary transfer helpers, camera-parameter dialogs, add/drop-file handling,
and the per-format initial document loader move into their own modules
behind thin host interfaces. 3,498 -> 1,245 lines. Also drops ~230 lines
of unreachable commented-out dialog code in handleCameraParametersRequest.

No behavior change: all 59 extension-host unit tests pass and the
webpack build is clean.
