+++ src/pointCloudEditorProvider.ts
@@ -10,27 +10,21 @@ import { PcdParser } from '../engine/src/parsers/pcdParser';
 import { PtsParser } from '../engine/src/parsers/ptsParser';
 import { OffParser } from '../engine/src/parsers/offParser';
 import { GltfParser } from '../engine/src/parsers/gltfParser';
-import { NpyParser } from '../engine/src/parsers/npyParser';
-import { XyzVariantParser } from '../engine/src/parsers/xyzVariantParser';
+import { sendUltimateRawBinary } from './providerHandlers/binaryTransfer';
 import {
-  parseXyzWasm,
-  parseAsciiPlyWasm,
-  parsePcdAsciiWasm,
-  parsePcdBinaryWasm,
-  parsePtsWasm,
-  streamParseFile,
-  detectXyzColorMode,
-} from './wasmPointcloud';
+  handleCameraParametersRequest,
+  handleCameraParametersWithScaleRequest,
+} from './providerHandlers/cameraParamsHandlers';
+import {
+  handleAddFile,
+  handleAddFileFromPath,
+  handleDroppedFilesFromWebview,
+  type AddFileHost,
+} from './providerHandlers/addFileHandlers';
+import { loadDocumentContent, type DocumentLoaderHost } from './providerHandlers/documentLoader';
 
 // Shared file handling functionality
-import {
-  detectFileType,
-  detectFileTypeWithContent,
-  isPlyBinary,
-  generateErrorMessage,
-  SUPPORTED_EXTENSIONS,
-  ALL_SUPPORTED_EXTENSIONS,
-} from '../engine/src/fileHandler';
+import { detectFileType, detectFileTypeWithContent, isPlyBinary } from '../engine/src/fileHandler';
 
 export class PointCloudEditorProvider implements vscode.CustomReadonlyEditorProvider {
   private static readonly viewType = 'plyViewer.plyEditor';
@@ -44,6 +38,23 @@ export class PointCloudEditorProvider implements vscode.CustomReadonlyEditorProv
   // every outgoing *Data message so the webview can report one consistent
   // end-to-end timing line (read+parse / transfer / build / total).
   private currentLoadStartedAt = 0;
+  private readonly addFileHost: AddFileHost = {
+    getShortPath: filePath => this.getShortPath(filePath),
+    logPerf: line => this.logPerf(line),
+    setLoadStartedAt: ts => {
+      this.currentLoadStartedAt = ts;
+    },
+    tryAutoLoadMtl: (webviewPanel, objUri, parsedObjData, fileIndex) =>
+      this.tryAutoLoadMtl(webviewPanel, objUri, parsedObjData, fileIndex),
+  };
+  private readonly documentLoaderHost: DocumentLoaderHost = {
+    getShortPath: filePath => this.getShortPath(filePath),
+    logPerf: line => this.logPerf(line),
+    getCurrentLoadStartedAt: () => this.currentLoadStartedAt,
+    tryAutoLoadMtl: (webviewPanel, objUri, parsedObjData, fileIndex) =>
+      this.tryAutoLoadMtl(webviewPanel, objUri, parsedObjData, fileIndex),
+    getSceneMetadata: fsPath => this.datasetManager.getSceneMetadata(fsPath),
+  };
 
   constructor(private readonly context: vscode.ExtensionContext) {
     this.datasetManager = new DatasetManager(context);
@@ -217,777 +228,25 @@ export class PointCloudEditorProvider implements vscode.CustomReadonlyEditorProv
     await this.handleRequestDefaultDepthSettings(webviewPanel);
 
     // Load and parse file asynchronously (don't await - let UI show first)
-    setImmediate(async () => {
-      try {
-        const loadStartTime = performance.now();
-        const wallStart = new Date().toISOString();
-
-        if (isDepthFile) {
-          // Handle depth files (TIF, PFM, NPY, NPZ, PNG, EXR) for point cloud conversion
-          const fileType = isPfmFile
-            ? 'PFM'
-            : isNpyFile
-              ? 'NPY'
-              : isPngFile
-                ? 'PNG'
-                : isExrFile
-                  ? 'EXR'
-                  : 'TIF';
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `🚀 Extension: Starting ${fileType} file processing for depth conversion...`,
-            timestamp: loadStartTime,
-          });
-
-          // Read depth file and send for webview processing
-          const depthData = await vscode.workspace.fs.readFile(document.uri);
-          const fileReadTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timing',
-            phase: 'read',
-            kind: 'depth',
-            ms: +(fileReadTime - loadStartTime).toFixed(1),
-          });
-
-          // Check for dataset scene metadata
-          const sceneMetadata = await this.datasetManager.getSceneMetadata(document.uri.fsPath);
-
-          // Send depth data to webview for conversion
-          webviewPanel.webview.postMessage({
-            type: 'depthData',
-            // Wall-clock epoch for measuring cross-process transfer cost in the webview.
-            postedAt: Date.now(),
-            fileName: path.basename(document.uri.fsPath),
-            shortPath: this.getShortPath(document.uri.fsPath),
-            data: depthData.buffer.slice(
-              depthData.byteOffset,
-              depthData.byteOffset + depthData.byteLength
-            ),
-            sceneMetadata: sceneMetadata || undefined,
-          });
-
-          return; // Exit early for depth files
-        }
-
-        if (isNpyPointCloud) {
-          // Handle NPY point cloud file (shape ending with 3)
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: '🚀 Extension: Starting NPY point cloud processing...',
-            timestamp: loadStartTime,
-          });
-          const npyData = await vscode.workspace.fs.readFile(document.uri);
-          const fileReadTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `📁 Extension: File read in ${(fileReadTime - loadStartTime).toFixed(2)}ms`,
-            timestamp: loadStartTime,
-          });
-
-          const npyParser = new NpyParser();
-          const parsedData = await npyParser.parse(npyData);
-
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `🔬 Extension: NPY parsing completed in ${performance.now() - fileReadTime}ms`,
-            timestamp: loadStartTime,
-          });
-
-          // Send parsed NPY data to webview
-          webviewPanel.webview.postMessage({
-            type: 'npyData',
-            fileName: path.basename(document.uri.fsPath),
-            shortPath: this.getShortPath(document.uri.fsPath),
-            data: parsedData,
-          });
-          return; // Exit early for NPY point cloud files
-        }
-
-        if (isObjFile) {
-          // Handle OBJ file
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: '🚀 Extension: Starting OBJ file processing...',
-            timestamp: loadStartTime,
-          });
-
-          const objData = await vscode.workspace.fs.readFile(document.uri);
-          const fileReadTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `📁 Extension: OBJ file read took ${(fileReadTime - loadStartTime).toFixed(1)}ms`,
-            timestamp: fileReadTime,
-          });
-
-          const objParser = new ObjParser();
-          const timingCallback = (message: string) => {
-            webviewPanel.webview.postMessage({
-              type: 'timingUpdate',
-              message: message,
-              timestamp: performance.now(),
-            });
-          };
-
-          const parsedData = await objParser.parse(objData, timingCallback);
-          const parseTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `🎯 Extension: OBJ parsing took ${(parseTime - fileReadTime).toFixed(1)}ms`,
-            timestamp: parseTime,
-          });
-
-          // Send parsed OBJ data to webview
-          webviewPanel.webview.postMessage({
-            type: 'objData',
-            fileName: path.basename(document.uri.fsPath),
-            shortPath: this.getShortPath(document.uri.fsPath),
-            fileSizeInBytes: objData.byteLength,
-            data: parsedData,
-          });
-
-          // Try to auto-load MTL file
-          await this.tryAutoLoadMtl(webviewPanel, document.uri, parsedData, 0);
-
-          return; // Exit early for OBJ files
-        }
-
-        if (isStlFile) {
-          // Handle STL file
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: '🚀 Extension: Starting STL file processing...',
-            timestamp: loadStartTime,
-          });
-
-          const stlData = await vscode.workspace.fs.readFile(document.uri);
-          const fileReadTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `📁 Extension: STL file read took ${(fileReadTime - loadStartTime).toFixed(1)}ms`,
-            timestamp: fileReadTime,
-          });
-
-          const stlParser = new StlParser();
-          const timingCallback = (message: string) => {
-            webviewPanel.webview.postMessage({
-              type: 'timingUpdate',
-              message: message,
-              timestamp: performance.now(),
-            });
-          };
-
-          const parsedData = await stlParser.parse(stlData, timingCallback);
-          const parseTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `🎯 Extension: STL parsing took ${(parseTime - fileReadTime).toFixed(1)}ms`,
-            timestamp: parseTime,
-          });
-
-          // Send parsed STL data to webview
-          webviewPanel.webview.postMessage({
-            type: 'stlData',
-            fileName: path.basename(document.uri.fsPath),
-            shortPath: this.getShortPath(document.uri.fsPath),
-            fileSizeInBytes: stlData.byteLength,
-            data: parsedData,
-          });
-
-          return; // Exit early for STL files
-        }
-
-        if (isPcdFile) {
-          // Handle PCD file
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: '🚀 Extension: Starting PCD file processing...',
-            timestamp: loadStartTime,
-          });
-
-          // Streaming overlap for ASCII PCD (same cold-cache win as XYZ: the next
-          // chunk's disk read overlaps the current chunk's parse). Gate on the
-          // HEADER only so we don't read the whole file first: require ASCII data
-          // and an identity VIEWPOINT (the WASM stream parser carries no viewpoint
-          // transform). Anything else falls through to the whole-file path below.
-          if (document.uri.scheme === 'file') {
-            try {
-              const head = await this.readFileHead(document.uri, 65536);
-              const headText = Buffer.from(head).toString('latin1');
-              const isAsciiPcd = /[\r\n]DATA\s+ascii/i.test('\n' + headText);
-              if (isAsciiPcd && this.pcdViewpointIsIdentity(head)) {
-                const streamed = await streamParseFile(document.uri.fsPath, 'pcd');
-                if (streamed) {
-                  let pcdBytes = head.byteLength;
-                  try {
-                    pcdBytes = fs.statSync(document.uri.fsPath).size;
-                  } catch {
-                    /* size is best-effort for logging only */
-                  }
-                  this.logPerf(
-                    `⏱️ PERF[pcd/ext] load ${(performance.now() - loadStartTime).toFixed(1)}ms (${streamed.vertexCount} pts, wasm-stream) for ${path.basename(document.uri.fsPath)}`
-                  );
-                  webviewPanel.webview.postMessage({
-                    type: 'xyzVariantData',
-                    fileName: path.basename(document.uri.fsPath),
-                    shortPath: this.getShortPath(document.uri.fsPath),
-                    fileSizeInBytes: pcdBytes,
-                    data: streamed,
-                    variant: 'pcd',
-                    parseMode: 'wasm-stream',
-                  });
-                  return;
-                }
-              }
-            } catch {
-              /* fall through to the whole-file path */
-            }
-          }
-
-          const pcdData = await this.readFileFast(document.uri);
-          const fileReadTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `📁 Extension: PCD file read took ${(fileReadTime - loadStartTime).toFixed(1)}ms`,
-            timestamp: fileReadTime,
-          });
-
-          // Fast path: Rust/WASM for ASCII PCD point clouds with an identity
-          // viewpoint. parse_pcd_ascii returns null for binary PCD; the
-          // viewpoint guard keeps clouds that need the VIEWPOINT transform on
-          // the JS path. Anything else falls through to the JS parser below.
-          if (this.pcdViewpointIsIdentity(pcdData)) {
-            // ASCII via WASM, then binary via WASM (binary PCD otherwise falls to
-            // the slow JS parser — ~11x slower). Both gated on identity viewpoint
-            // since the WASM parsers don't carry the VIEWPOINT transform.
-            const pcdWasm = parsePcdAsciiWasm(pcdData) || parsePcdBinaryWasm(pcdData);
-            if (pcdWasm) {
-              this.logPerf(
-                `⏱️ PERF[pcd/ext] parse ${(performance.now() - fileReadTime).toFixed(1)}ms (${pcdWasm.vertexCount} pts, wasm) for ${path.basename(document.uri.fsPath)}`
-              );
-              webviewPanel.webview.postMessage({
-                type: 'xyzVariantData',
-                fileName: path.basename(document.uri.fsPath),
-                shortPath: this.getShortPath(document.uri.fsPath),
-                fileSizeInBytes: pcdData.byteLength,
-                data: pcdWasm,
-                variant: 'pcd',
-                parseMode: 'wasm',
-              });
-              return;
-            }
-          }
-
-          const pcdParser = new PcdParser();
-          const timingCallback = (message: string) => {
-            webviewPanel.webview.postMessage({
-              type: 'timingUpdate',
-              message: message,
-              timestamp: performance.now(),
-            });
-          };
-
-          const parsedData = await pcdParser.parse(pcdData, timingCallback);
-          const parseTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `🎯 Extension: PCD parsing took ${(parseTime - fileReadTime).toFixed(1)}ms`,
-            timestamp: parseTime,
-          });
-
-          // Send parsed PCD data to webview
-          webviewPanel.webview.postMessage({
-            type: 'pcdData',
-            fileName: path.basename(document.uri.fsPath),
-            shortPath: this.getShortPath(document.uri.fsPath),
-            fileSizeInBytes: pcdData.byteLength,
-            data: parsedData,
-          });
-
-          return; // Exit early for PCD files
-        }
-
-        if (isPtsFile) {
-          // Handle PTS file
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: '🚀 Extension: Starting PTS file processing...',
-            timestamp: loadStartTime,
-          });
-
-          const ptsData = await this.readFileFast(document.uri);
-          const fileReadTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `📁 Extension: PTS file read took ${(fileReadTime - loadStartTime).toFixed(1)}ms`,
-            timestamp: fileReadTime,
-          });
-
-          // Try the Rust/WASM parser (~2.5-3x faster); fall back to JS.
-          let parsedData: any;
-          let ptsMode = 'js';
-          const ptsWasm = parsePtsWasm(ptsData);
-          if (ptsWasm) {
-            parsedData = {
-              vertexCount: ptsWasm.vertexCount,
-              positionsArray: ptsWasm.positionsArray,
-              colorsArray: ptsWasm.colorsArray,
-              normalsArray: ptsWasm.normalsArray,
-              intensityArray: ptsWasm.intensityArray,
-              hasColors: ptsWasm.hasColors,
-              hasNormals: ptsWasm.hasNormals,
-              hasIntensity: ptsWasm.hasIntensity,
-              scalarFields: ptsWasm.intensityArray ? { intensity: ptsWasm.intensityArray } : {},
-              detectedFormat: `x y z${ptsWasm.hasIntensity ? ' intensity' : ''}${ptsWasm.hasColors ? ' r g b' : ''}`,
-              comments: [],
-            };
-            ptsMode = 'wasm';
-          } else {
-            const ptsParser = new PtsParser();
-            parsedData = await ptsParser.parse(ptsData);
-          }
-          this.logPerf(
-            `⏱️ PERF[pts/ext] parse ${(performance.now() - fileReadTime).toFixed(1)}ms (${parsedData.vertexCount} pts, ${ptsMode}) for ${path.basename(document.uri.fsPath)}`
-          );
-
-          // Send parsed PTS data to webview
-          webviewPanel.webview.postMessage({
-            type: 'ptsData',
-            fileName: path.basename(document.uri.fsPath),
-            shortPath: this.getShortPath(document.uri.fsPath),
-            fileSizeInBytes: ptsData.byteLength,
-            data: parsedData,
-            parseMode: ptsMode,
-          });
-
-          return; // Exit early for PTS files
-        }
-
-        if (isOffFile) {
-          // Handle OFF file
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: '🚀 Extension: Starting OFF file processing...',
-            timestamp: loadStartTime,
-          });
-
-          const offData = await vscode.workspace.fs.readFile(document.uri);
-          const fileReadTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `📁 Extension: OFF file read took ${(fileReadTime - loadStartTime).toFixed(1)}ms`,
-            timestamp: fileReadTime,
-          });
-
-          const offParser = new OffParser();
-          const timingCallback = (message: string) => {
-            webviewPanel.webview.postMessage({
-              type: 'timingUpdate',
-              message: message,
-              timestamp: performance.now(),
-            });
-          };
-
-          const parsedData = await offParser.parse(offData, timingCallback);
-          const parseTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `🎯 Extension: OFF parsing took ${(parseTime - fileReadTime).toFixed(1)}ms`,
-            timestamp: parseTime,
-          });
-
-          // Send parsed OFF data to webview
-          webviewPanel.webview.postMessage({
-            type: 'offData',
-            fileName: path.basename(document.uri.fsPath),
-            shortPath: this.getShortPath(document.uri.fsPath),
-            fileSizeInBytes: offData.byteLength,
-            data: parsedData,
-          });
-
-          return; // Exit early for OFF files
-        }
-
-        if (isGltfFile) {
-          // Handle GLTF/GLB file
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: '🚀 Extension: Starting GLTF/GLB file processing...',
-            timestamp: loadStartTime,
-          });
-
-          const gltfData = await vscode.workspace.fs.readFile(document.uri);
-          const fileReadTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `📁 Extension: GLTF/GLB file read took ${(fileReadTime - loadStartTime).toFixed(1)}ms`,
-            timestamp: fileReadTime,
-          });
-
-          const gltfParser = new GltfParser();
-          const timingCallback = (message: string) => {
-            webviewPanel.webview.postMessage({
-              type: 'timingUpdate',
-              message: message,
-              timestamp: performance.now(),
-            });
-          };
-
-          const parsedData = await gltfParser.parse(gltfData, timingCallback);
-          const parseTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `🎯 Extension: GLTF/GLB parsing took ${(parseTime - fileReadTime).toFixed(1)}ms`,
-            timestamp: parseTime,
-          });
-
-          // Send parsed GLTF/GLB data to webview
-          webviewPanel.webview.postMessage({
-            type: 'gltfData',
-            fileName: path.basename(document.uri.fsPath),
-            shortPath: this.getShortPath(document.uri.fsPath),
-            fileSizeInBytes: gltfData.byteLength,
-            data: parsedData,
-          });
-
-          return; // Exit early for GLTF/GLB files
-        }
-
-        if (isXyzVariant) {
-          // Handle XYZN/XYZRGB variants - send to webview for processing
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: '🚀 Extension: Starting XYZ variant file processing...',
-            timestamp: loadStartTime,
-          });
-
-          const xyzVariant =
-            fileType?.extension === 'xyzn'
-              ? 'xyzn'
-              : fileType?.extension === 'xyzrgb'
-                ? 'xyzrgb'
-                : 'xyz';
-
-          // Streaming overlap (Rust/WASM StreamParser): the next chunk's disk
-          // read (libuv) runs while the current chunk parses (main thread), so
-          // total ≈ max(read, parse). A controlled cold A/B (purge between each)
-          // showed it saves ~0.6–1.2s on 600MB XYZ files — cold is the
-          // real-world first-open case. It costs ~300ms only when re-opening an
-          // already-cached file (warm), which is rare. So: on for local files.
-          const ENABLE_XYZ_STREAMING = true;
-          const loadStart = performance.now();
-          let xyzParsed: any = null;
-          let xyzMode = 'js';
-          let xyzBytes = 0;
-
-          // XYZRGB has no type header, so decide once per file whether its colors
-          // are 0-255 ints or 0-1 floats by checking the color tokens' text (a
-          // decimal point ⇒ float). Done from a small header sample.
-          let xyzColorMode = 'auto';
-          if (xyzVariant === 'xyzrgb' && document.uri.scheme === 'file') {
-            try {
-              const head = await this.readFileHead(document.uri, 65536);
-              xyzColorMode = detectXyzColorMode(head, xyzVariant);
-            } catch {
-              /* fall back to the in-parser value heuristic */
-            }
-          }
-
-          if (ENABLE_XYZ_STREAMING && document.uri.scheme === 'file') {
-            const streamed = await streamParseFile(document.uri.fsPath, xyzVariant, xyzColorMode);
-            if (streamed) {
-              xyzParsed = streamed;
-              xyzMode = 'wasm-stream';
-              try {
-                xyzBytes = fs.statSync(document.uri.fsPath).size;
-              } catch {
-                /* size is best-effort for logging only */
-              }
-            }
-          }
-          if (!xyzParsed) {
-            const xyzData = await this.readFileFast(document.uri);
-            xyzBytes = xyzData.byteLength;
-            const wasmParsed = parseXyzWasm(xyzData, xyzVariant, xyzColorMode);
-            xyzParsed = wasmParsed ?? new XyzVariantParser().parse(xyzData, xyzVariant);
-            xyzMode = wasmParsed ? 'wasm' : 'js';
-          }
-          this.logPerf(
-            `⏱️ PERF[xyz/ext] load ${(performance.now() - loadStart).toFixed(1)}ms (${xyzParsed.vertexCount} pts, ${xyzMode}) for ${path.basename(document.uri.fsPath)}`
-          );
-
-          webviewPanel.webview.postMessage({
-            type: 'xyzVariantData',
-            fileName: path.basename(document.uri.fsPath),
-            shortPath: this.getShortPath(document.uri.fsPath),
-            fileSizeInBytes: xyzBytes,
-            data: xyzParsed,
-            variant: xyzVariant,
-            parseMode: xyzMode,
-          });
-
-          return; // Exit early for XYZ variant files
-        }
-
-        // Handle JSON pose files
-        if (isJsonFile) {
-          try {
-            const jsonBytes = await vscode.workspace.fs.readFile(document.uri);
-            const jsonText = Buffer.from(jsonBytes).toString('utf-8');
-
-            // Try standard parse first
-            let parsed: any;
-            try {
-              parsed = JSON.parse(jsonText);
-            } catch (e) {
-              // Fallback: sanitize non-standard tokens (NaN, Infinity) often found in pose dumps
-              const sanitizedText = jsonText
-                .replace(/\bNaN\b/g, 'null')
-                .replace(/\bInfinity\b/g, 'null')
-                .replace(/\b-Infinity\b/g, 'null');
-              parsed = JSON.parse(sanitizedText);
-            }
-
-            webviewPanel.webview.postMessage({
-              type: 'poseData',
-              fileName: path.basename(document.uri.fsPath),
-              shortPath: this.getShortPath(document.uri.fsPath),
-              data: parsed,
-            });
-            return; // Exit early for JSON pose files
-          } catch (err) {
-            webviewPanel.webview.postMessage({
-              type: 'loadingError',
-              fileName: path.basename(document.uri.fsPath),
-              fileType: 'JSON',
-              error: err instanceof Error ? err.message : String(err),
-            });
-            return;
-          }
-        }
-
-        // Send timing updates to webview for visibility
-        webviewPanel.webview.postMessage({
-          type: 'timing',
-          phase: 'start',
-          kind: 'ply',
-          at: wallStart,
-        });
-
-        // Streaming overlap for ASCII PLY point clouds (same cold-cache win as
-        // XYZ). Gate on a 64KB header read so we don't read the whole file first:
-        // only ASCII PLY streams (binary PLY uses the transfer-via-fetch path
-        // below); the stream parser returns null for binary/meshes → falls
-        // through to the whole-file path unchanged.
-        if (document.uri.scheme === 'file') {
-          try {
-            const head = await this.readFileHead(document.uri, 65536);
-            if (!isPlyBinary(head)) {
-              const streamed = await streamParseFile(document.uri.fsPath, 'ply');
-              if (streamed) {
-                let plyBytes = head.byteLength;
-                try {
-                  plyBytes = fs.statSync(document.uri.fsPath).size;
-                } catch {
-                  /* size is best-effort for logging only */
-                }
-                this.logPerf(
-                  `⏱️ PERF[ply/ext] load ${(performance.now() - loadStartTime).toFixed(1)}ms (${streamed.vertexCount} pts, wasm-stream) for ${path.basename(document.uri.fsPath)}`
-                );
-                webviewPanel.webview.postMessage({
-                  type: 'xyzVariantData',
-                  fileName: path.basename(document.uri.fsPath),
-                  shortPath: this.getShortPath(document.uri.fsPath),
-                  fileSizeInBytes: plyBytes,
-                  data: streamed,
-                  variant: 'ply',
-                  parseMode: 'wasm-stream',
-                });
-                return;
-              }
-            }
-          } catch {
-            /* fall through to the whole-file path */
-          }
-        }
-
-        const spatialData = await vscode.workspace.fs.readFile(document.uri);
-        const fileReadTime = performance.now();
-        webviewPanel.webview.postMessage({
-          type: 'timing',
-          phase: 'read',
-          kind: 'ply',
-          ms: +(fileReadTime - loadStartTime).toFixed(1),
-        });
-
-        const parser = new PlyParser();
-        webviewPanel.webview.postMessage({
-          type: 'timingUpdate',
-          message: '🚀 Extension: ULTIMATE - Starting header-only parsing...',
-          timestamp: performance.now(),
-        });
-
-        // Create timing callback that forwards to webview
-        const timingCallback = (message: string) => {
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: message,
-            timestamp: performance.now(),
-          });
-        };
-
-        // Detect format first using shared functionality
-        const isBinary = isPlyBinary(spatialData);
-
-        if (isBinary) {
-          // Binary PLY - use ULTIMATE parsing
-          const headerResult = await parser.parseHeaderOnly(spatialData, timingCallback);
-          const parsedData = headerResult.headerInfo;
-          const parseTime = performance.now();
-          webviewPanel.webview.postMessage({
-            type: 'timing',
-            phase: 'parse',
-            kind: 'ply',
-            format: parsedData.format,
-            ms: +(parseTime - fileReadTime).toFixed(1),
-          });
-
-          // Add file info
-          parsedData.fileName = path.basename(document.uri.fsPath);
-          parsedData.shortPath = this.getShortPath(document.uri.fsPath);
-          parsedData.fileIndex = 0;
-
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: '🚀 Extension: Starting binary data conversion...',
-            timestamp: performance.now(),
-          });
-
-          // ULTIMATE: Send raw binary data for webview-side parsing
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: '🚀 Extension: ULTIMATE - Sending raw binary data...',
-            timestamp: performance.now(),
-          });
-
-          // Send raw binary data + header info
-          // Extra logging to aid debugging face offsets/types
-          // Log face types once for debugging
-          // concise header info for debugging (once)
-          webviewPanel.webview.postMessage({
-            type: 'timingUpdate',
-            message: `Header face types: count=${headerResult.faceCountType || 'n/a'}, index=${headerResult.faceIndexType || 'n/a'}`,
-            timestamp: performance.now(),
-          });
-          // Transfer-via-fetch: send only header metadata + a webview URI for
-          // the file. The webview fetches the bytes directly, avoiding the
-          // multi-hundred-ms structured-clone of the full vertex buffer. On a
-          // fetch failure the webview asks the extension to resend over
-          // postMessage (see the 'plyFetchFailed' handler, which re-reads).
-          webviewPanel.webview.postMessage({
-            type: 'ultimateRawBinaryUri',
-            messageType: 'multiSpatialData',
-            postedAt: Date.now(),
-            loadStartedAt: this.currentLoadStartedAt,
-            fileUri: webviewPanel.webview.asWebviewUri(document.uri).toString(),
-            docUri: document.uri.toString(),
-            binaryDataStart: headerResult.binaryDataStart,
-            fileName: parsedData.fileName,
-            shortPath: parsedData.shortPath,
-            fileSizeInBytes: spatialData.byteLength,
-            vertexCount: parsedData.vertexCount,
-            faceCount: parsedData.faceCount,
-            hasColors: parsedData.hasColors,
-            hasNormals: parsedData.hasNormals,
-            hasIntensity: parsedData.hasIntensity,
-            format: parsedData.format,
-            comments: parsedData.comments,
-            vertexStride: headerResult.vertexStride,
-            propertyOffsets: Array.from(headerResult.propertyOffsets.entries()),
-            littleEndian: headerResult.headerInfo.format === 'binary_little_endian',
-            faceCountType: headerResult.faceCountType,
-            faceIndexType: headerResult.faceIndexType,
-          });
-        } else {
-          // ASCII PLY. Try the Rust/WASM parser first — it handles point clouds
-          // (no faces) and returns null for meshes or on any failure, so those
-          // transparently fall through to the JS parser below.
-          const plyWasm = parseAsciiPlyWasm(spatialData);
-          if (plyWasm) {
-            this.logPerf(
-              `⏱️ PERF[ply/ext] parse ${(performance.now() - fileReadTime).toFixed(1)}ms (${plyWasm.vertexCount} pts, wasm) for ${path.basename(document.uri.fsPath)}`
-            );
-            webviewPanel.webview.postMessage({
-              type: 'xyzVariantData',
-              fileName: path.basename(document.uri.fsPath),
-              shortPath: this.getShortPath(document.uri.fsPath),
-              fileSizeInBytes: spatialData.byteLength,
-              data: plyWasm,
-              variant: 'ply',
-              parseMode: 'wasm',
-            });
-            return; // handled by the fast path
-          }
-          console.log(
-            `📝 ASCII PLY detected: ${path.basename(document.uri.fsPath)} - using traditional parsing`
-          );
-          const parsedData = await parser.parse(spatialData, timingCallback);
-          const parseTime = performance.now();
-          const isXyz = /\.xyz$/i.test(document.uri.fsPath);
-          this.logPerf(
-            `⏱️ PERF[${isXyz ? 'xyz' : 'ply'}/ext] parse ${(parseTime - fileReadTime).toFixed(1)}ms (${parsedData.vertexCount} pts) for ${path.basename(document.uri.fsPath)}`
-          );
-          webviewPanel.webview.postMessage({
-            type: 'timing',
-            phase: 'parse',
-            kind: 'ply',
-            format: parsedData.format,
-            ms: +(parseTime - fileReadTime).toFixed(1),
-          });
-
-          // Add file info
-          parsedData.fileName = path.basename(document.uri.fsPath);
-          parsedData.shortPath = this.getShortPath(document.uri.fsPath);
-          parsedData.fileIndex = 0;
-          (parsedData as any).fileSizeInBytes = spatialData.byteLength;
-
-          // Send via traditional method (will use binary transfer if possible)
-          await this.sendSpatialDataToWebview(webviewPanel, [parsedData], 'multiSpatialData');
-        }
-        const totalTime = performance.now();
-        webviewPanel.webview.postMessage({
-          type: 'timing',
-          phase: 'total',
-          kind: 'ply',
-          ms: +(totalTime - loadStartTime).toFixed(1),
-          at: new Date().toISOString(),
-        });
-      } catch (error) {
-        console.error(`Extension: PLY processing failed:`, error);
-        const ext = path.extname(document.uri.fsPath).toLowerCase();
-        const fileTypeMap: Record<string, string> = {
-          '.ply': 'PLY',
-          '.pcd': 'PCD',
-          '.stl': 'STL',
-          '.obj': 'OBJ',
-          '.xyz': 'XYZ',
-          '.pts': 'PTS',
-          '.tif': 'TIFF',
-          '.tiff': 'TIFF',
-          '.pfm': 'PFM',
-          '.npy': 'NPY',
-          '.npz': 'NPZ',
-          '.png': 'PNG',
-          '.json': 'JSON',
-        };
-        const fileType = fileTypeMap[ext] || ext.toUpperCase().slice(1) || 'point cloud';
-        webviewPanel.webview.postMessage({
-          type: 'loadingError',
-          fileName: path.basename(document.uri.fsPath),
-          fileType,
-          error: error instanceof Error ? error.message : String(error),
-        });
-      }
-    });
+    setImmediate(() =>
+      loadDocumentContent(this.documentLoaderHost, document.uri, webviewPanel, {
+        fileType,
+        isDepthFile,
+        isPfmFile,
+        isNpyFile,
+        isPngFile,
+        isExrFile,
+        isNpyPointCloud,
+        isObjFile,
+        isStlFile,
+        isPcdFile,
+        isPtsFile,
+        isOffFile,
+        isGltfFile,
+        isXyzVariant,
+        isJsonFile,
+      })
+    );
 
     // Handle messages from webview
     webviewPanel.webview.onDidReceiveMessage(async message => {
@@ -1005,7 +264,7 @@ export class PointCloudEditorProvider implements vscode.CustomReadonlyEditorProv
           await this.handlePlyFetchFallback(message);
           break;
         case 'addFile':
-          await this.handleAddFile(webviewPanel);
+          await handleAddFile(this.addFileHost, webviewPanel, this.panelToPath.get(webviewPanel));
           break;
         case 'removeFile':
           webviewPanel.webview.postMessage({
@@ -1015,11 +274,11 @@ export class PointCloudEditorProvider implements vscode.CustomReadonlyEditorProv
           break;
         case 'requestCameraParams':
           // Request camera parameters for TIF conversion
-          await this.handleCameraParametersRequest(webviewPanel, message);
+          await handleCameraParametersRequest(this.context, webviewPanel, message);
           break;
         case 'requestCameraParamsWithScale':
           // Request camera parameters with scale factor for PNG conversion
-          await this.handleCameraParametersWithScaleRequest(webviewPanel, message);
+          await handleCameraParametersWithScaleRequest(this.context, webviewPanel, message);
           break;
         case 'savePlyFile':
           // Handle PLY file save request
@@ -1047,10 +306,10 @@ export class PointCloudEditorProvider implements vscode.CustomReadonlyEditorProv
           await this.handleSequenceRequestFile(webviewPanel, message);
           break;
         case 'addFileFromPath':
-          await this.handleAddFileFromPath(webviewPanel, message.path as string);
+          await handleAddFileFromPath(this.addFileHost, webviewPanel, message.path as string);
           break;
         case 'addDroppedFiles':
-          await this.handleDroppedFilesFromWebview(webviewPanel, message.files || []);
+          await handleDroppedFilesFromWebview(this.addFileHost, webviewPanel, message.files || []);
           break;
         case 'requestDatasetTexture':
           await this.handleRequestDatasetTexture(webviewPanel, message);
@@ -1193,747 +452,6 @@ export class PointCloudEditorProvider implements vscode.CustomReadonlyEditorProv
     html = html.replace(/<div class="bottom-right-nav">[\s\S]*?<\/div>/, '');
 
     return html;
-  }
-
-  private async handleAddFile(webviewPanel: vscode.WebviewPanel): Promise<void> {
-    // Get current file directory for default location
-    const currentFilePath = this.panelToPath.get(webviewPanel);
-    const defaultUri = currentFilePath ? vscode.Uri.file(path.dirname(currentFilePath)) : undefined;
-
-    const files = await vscode.window.showOpenDialog({
-      canSelectMany: true,
-      filters: {
-        'Point Cloud & Pose Files': [...ALL_SUPPORTED_EXTENSIONS],
-
-        'Point Clouds': [...SUPPORTED_EXTENSIONS.pointClouds],
-
-        Meshes: [...SUPPORTED_EXTENSIONS.meshes],
-
-        'Depth Images': [...SUPPORTED_EXTENSIONS.depthImages],
-
-        'Pose Data': [...SUPPORTED_EXTENSIONS.poseData],
-      },
-      title: 'Select point cloud files to add',
-      defaultUri: defaultUri,
-    });
-
-    if (files && files.length > 0) {
-      for (let i = 0; i < files.length; i++) {
-        try {
-          const fileStartTime = performance.now();
-          const fileName = path.basename(files[i].fsPath);
-          const shortPath = this.getShortPath(files[i].fsPath);
-          const fileExtension = path.extname(files[i].fsPath).toLowerCase();
-          console.log(`🚀 ULTIMATE: Processing add file ${fileName} (${fileExtension})`);
-
-          // Tell the webview a load started so it shows a non-blocking progress
-          // row in the Files list (the scene already has clouds and stays
-          // interactive — no overlay). Sent before the read/parse below.
-          this.currentLoadStartedAt = Date.now();
-          webviewPanel.webview.postMessage({ type: 'startLoading', fileName });
-
-          // Handle different file types
-          if (
-            fileExtension === '.tif' ||
-            fileExtension === '.tiff' ||
-            fileExtension === '.pfm' ||
-            fileExtension === '.npy' ||
-            fileExtension === '.npz' ||
-            fileExtension === '.png' ||
-            fileExtension === '.exr'
-          ) {
-            // Handle depth files for conversion
-            const depthData = await vscode.workspace.fs.readFile(files[i]);
-
-            // Send depth data to webview for conversion
-            webviewPanel.webview.postMessage({
-              type: 'depthData',
-              fileName: fileName,
-              shortPath: shortPath,
-              data: depthData.buffer.slice(
-                depthData.byteOffset,
-                depthData.byteOffset + depthData.byteLength
-              ),
-              isAddFile: true, // Flag to indicate this is from "Add Point Cloud"
-            });
-
-            console.log(`🎯 Depth Add File: ${fileName} sent for processing`);
-            continue;
-          }
-
-          // Handle PLY files (existing logic)
-          if (fileExtension === '.ply') {
-            // Read file data
-            const spatialData = await vscode.workspace.fs.readFile(files[i]);
-            const fileReadTime = performance.now();
-
-            // Parse file (detect format first)
-            const parser = new PlyParser();
-
-            // Quick format detection using shared functionality
-            const isBinary = isPlyBinary(spatialData);
-
-            if (isBinary) {
-              // Use ultimate binary transfer for binary PLY files
-              const headerResult = await parser.parseHeaderOnly(spatialData);
-              const parseTime = performance.now();
-
-              // Add file info
-              headerResult.headerInfo.fileName = fileName;
-              headerResult.headerInfo.shortPath = shortPath;
-              headerResult.headerInfo.fileIndex = i;
-
-              // Send ultimate raw binary data
-              await this.sendUltimateRawBinary(
-                webviewPanel,
-                headerResult.headerInfo,
-                headerResult,
-                spatialData,
-                'addFiles'
-              );
-            } else {
-              // Use traditional parsing for ASCII PLY files
-              console.log(`📝 ASCII PLY detected: ${fileName} - using traditional parsing`);
-              const parsedData = await parser.parse(spatialData);
-              const parseTime = performance.now();
-
-              // Add file info
-              parsedData.fileName = fileName;
-              parsedData.shortPath = shortPath;
-              parsedData.fileIndex = i;
-
-              // Send via traditional method (will use binary transfer if possible)
-              await this.sendSpatialDataToWebview(webviewPanel, [parsedData], 'addFiles');
-            }
-
-            const totalTime = performance.now();
-            console.log(
-              `🎯 ULTIMATE Add PLY File: ${fileName} processed in ${(totalTime - fileStartTime).toFixed(1)}ms`
-            );
-            continue;
-          }
-
-          // Handle XYZ files
-          if (fileExtension === '.xyz') {
-            const xyzData = await vscode.workspace.fs.readFile(files[i]);
-
-            // Send XYZ data to webview for parsing
-            webviewPanel.webview.postMessage({
-              type: 'xyzData',
-              fileName: fileName,
-              shortPath: shortPath,
-              data: xyzData.buffer.slice(
-                xyzData.byteOffset,
-                xyzData.byteOffset + xyzData.byteLength
-              ),
-              isAddFile: true,
-            });
-
-            console.log(`🎯 XYZ Add File: ${fileName} sent for processing`);
-            continue;
-          }
-
-          // Handle OBJ files
-          if (fileExtension === '.obj') {
-            const objData = await vscode.workspace.fs.readFile(files[i]);
-            const objParser = new ObjParser();
-            const parsedData = await objParser.parse(objData);
-
-            webviewPanel.webview.postMessage({
-              type: 'objData',
-              fileName: fileName,
-              shortPath: shortPath,
-              data: parsedData,
-              isAddFile: true,
-            });
-
-            // Try to auto-load MTL file for added OBJ files
-            await this.tryAutoLoadMtl(webviewPanel, files[i], parsedData, i);
-
-            console.log(`🎯 OBJ Add File: ${fileName} sent for processing`);
-            continue;
-          }
-
-          // Handle STL files
-          if (fileExtension === '.stl') {
-            const stlData = await vscode.workspace.fs.readFile(files[i]);
-            const stlParser = new StlParser();
-            const parsedData = await stlParser.parse(stlData);
-
-            webviewPanel.webview.postMessage({
-              type: 'stlData',
-              fileName: fileName,
-              shortPath: shortPath,
-              data: parsedData,
-              isAddFile: true,
-            });
-
-            console.log(`🎯 STL Add File: ${fileName} sent for processing`);
-            continue;
-          }
-
-          // Handle JSON pose files
-          if (fileExtension === '.json') {
-            try {
-              const jsonBytes = await vscode.workspace.fs.readFile(files[i]);
-              const jsonText = Buffer.from(jsonBytes).toString('utf-8');
-              // Try standard parse first, then sanitize fallback
-              let parsed: any;
-              try {
-                parsed = JSON.parse(jsonText);
-              } catch (e) {
-                const sanitizedText = jsonText
-                  .replace(/\bNaN\b/g, 'null')
-                  .replace(/\bInfinity\b/g, 'null')
-                  .replace(/\b-Infinity\b/g, 'null');
-                parsed = JSON.parse(sanitizedText);
-              }
-              webviewPanel.webview.postMessage({
-                type: 'poseData',
-                fileName: fileName,
-                shortPath: shortPath,
-                data: parsed,
-                isAddFile: true,
-              });
-              console.log(`🎯 JSON Pose Add File: ${fileName} sent for processing`);
-            } catch (err) {
-              vscode.window.showErrorMessage(
-                `Failed to load JSON pose ${fileName}: ${err instanceof Error ? err.message : String(err)}`
-              );
-            }
-            continue;
-          }
-
-          // Handle PCD files
-          if (fileExtension === '.pcd') {
-            const pcdData = await vscode.workspace.fs.readFile(files[i]);
-            const pcdParser = new PcdParser();
-            const parsedData = await pcdParser.parse(pcdData);
-
-            webviewPanel.webview.postMessage({
-              type: 'pcdData',
-              fileName: fileName,
-              shortPath: shortPath,
-              data: parsedData,
-              isAddFile: true,
-            });
-
-            console.log(`🎯 PCD Add File: ${fileName} sent for processing`);
-            continue;
-          }
-
-          // Handle PTS files
-          if (fileExtension === '.pts') {
-            const ptsData = await vscode.workspace.fs.readFile(files[i]);
-            const ptsParser = new PtsParser();
-            const parsedData = await ptsParser.parse(ptsData);
-
-            webviewPanel.webview.postMessage({
-              type: 'ptsData',
-              fileName: fileName,
-              shortPath: shortPath,
-              data: parsedData,
-              isAddFile: true,
-            });
-
-            console.log(`🎯 PTS Add File: ${fileName} sent for processing`);
-            continue;
-          }
-
-          // Handle OFF files
-          if (fileExtension === '.off') {
-            const offData = await vscode.workspace.fs.readFile(files[i]);
-            const offParser = new OffParser();
-            const parsedData = await offParser.parse(offData);
-
-            webviewPanel.webview.postMessage({
-              type: 'offData',
-              fileName: fileName,
-              shortPath: shortPath,
-              data: parsedData,
-              isAddFile: true,
-            });
-
-            console.log(`🎯 OFF Add File: ${fileName} sent for processing`);
-            continue;
-          }
-
-          // Handle GLTF/GLB files
-          if (fileExtension === '.gltf' || fileExtension === '.glb') {
-            const gltfData = await vscode.workspace.fs.readFile(files[i]);
-            const gltfParser = new GltfParser();
-            const parsedData = await gltfParser.parse(gltfData);
-
-            webviewPanel.webview.postMessage({
-              type: 'gltfData',
-              fileName: fileName,
-              shortPath: shortPath,
-              data: parsedData,
-              isAddFile: true,
-            });
-
-            console.log(`🎯 GLTF/GLB Add File: ${fileName} sent for processing`);
-            continue;
-          }
-
-          // Handle XYZN/XYZRGB variants
-          if (fileExtension === '.xyzn' || fileExtension === '.xyzrgb') {
-            const xyzData = await vscode.workspace.fs.readFile(files[i]);
-
-            webviewPanel.webview.postMessage({
-              type: 'xyzVariantData',
-              fileName: fileName,
-              shortPath: shortPath,
-              data: xyzData.buffer.slice(
-                xyzData.byteOffset,
-                xyzData.byteOffset + xyzData.byteLength
-              ),
-              variant: fileExtension.substring(1), // Remove the dot
-              isAddFile: true,
-            });
-
-            console.log(`🎯 XYZ Variant Add File: ${fileName} sent for processing`);
-            continue;
-          }
-
-          // Unsupported file type
-          vscode.window.showWarningMessage(
-            generateErrorMessage(
-              fileName,
-              fileExtension.substring(1),
-              new Error('Unsupported file type')
-            )
-          );
-        } catch (error) {
-          console.error(`Failed to load file ${files[i].fsPath}:`, error);
-          vscode.window.showErrorMessage(`Failed to load file ${files[i].fsPath}: ${error}`);
-        }
-      }
-    }
-  }
-
-  private async handleAddFileFromPath(
-    webviewPanel: vscode.WebviewPanel,
-    filePathStr: string
-  ): Promise<void> {
-    try {
-      const fileUri = vscode.Uri.file(filePathStr);
-      const fileName = path.basename(fileUri.fsPath);
-      const shortPath = this.getShortPath(fileUri.fsPath);
-      const ext = path.extname(fileUri.fsPath).toLowerCase();
-
-      // Non-blocking progress row in the Files list during read/parse (the scene
-      // already has clouds and stays interactive).
-      this.currentLoadStartedAt = Date.now();
-      webviewPanel.webview.postMessage({ type: 'startLoading', fileName });
-
-      if (
-        ext === '.tif' ||
-        ext === '.tiff' ||
-        ext === '.pfm' ||
-        ext === '.npy' ||
-        ext === '.npz' ||
-        ext === '.png' ||
-        ext === '.exr'
-      ) {
-        const depthData = await vscode.workspace.fs.readFile(fileUri);
-        webviewPanel.webview.postMessage({
-          type: 'depthData',
-          fileName,
-          shortPath,
-          data: depthData.buffer.slice(
-            depthData.byteOffset,
-            depthData.byteOffset + depthData.byteLength
-          ),
-          isAddFile: true,
-        });
-        return;
-      }
-      if (ext === '.ply') {
-        const spatialData = await vscode.workspace.fs.readFile(fileUri);
-        const parser = new PlyParser();
-        const isBinary = isPlyBinary(spatialData);
-        if (isBinary) {
-          const headerResult = await parser.parseHeaderOnly(spatialData);
-          headerResult.headerInfo.fileName = fileName;
-          headerResult.headerInfo.shortPath = shortPath;
-          await this.sendUltimateRawBinary(
-            webviewPanel,
-            headerResult.headerInfo,
-            headerResult,
-            spatialData,
-            'addFiles'
-          );
-        } else {
-          const parsedData = await parser.parse(spatialData);
-          parsedData.fileName = fileName;
-          parsedData.shortPath = shortPath;
-          await this.sendSpatialDataToWebview(webviewPanel, [parsedData], 'addFiles');
-        }
-        return;
-      }
-      if (ext === '.xyz') {
-        const xyzData = await vscode.workspace.fs.readFile(fileUri);
-        webviewPanel.webview.postMessage({
-          type: 'xyzData',
-          fileName,
-          shortPath,
-          data: xyzData.buffer.slice(xyzData.byteOffset, xyzData.byteOffset + xyzData.byteLength),
-          isAddFile: true,
-        });
-        return;
-      }
-      if (ext === '.obj') {
-        const objData = await vscode.workspace.fs.readFile(fileUri);
-        const objParser = new ObjParser();
-        const parsedData = await objParser.parse(objData);
-        webviewPanel.webview.postMessage({
-          type: 'objData',
-          fileName,
-          shortPath,
-          data: parsedData,
-          isAddFile: true,
-        });
-        await this.tryAutoLoadMtl(webviewPanel, fileUri, parsedData, 0);
-        return;
-      }
-      if (ext === '.stl') {
-        const stlData = await vscode.workspace.fs.readFile(fileUri);
-        const stlParser = new StlParser();
-        const parsedData = await stlParser.parse(stlData);
-        webviewPanel.webview.postMessage({
-          type: 'stlData',
-          fileName,
-          shortPath,
-          data: parsedData,
-          isAddFile: true,
-        });
-        return;
-      }
-      if (ext === '.pcd') {
-        const pcdData = await vscode.workspace.fs.readFile(fileUri);
-        const pcdParser = new PcdParser();
-        const parsedData = await pcdParser.parse(pcdData);
-        webviewPanel.webview.postMessage({
-          type: 'pcdData',
-          fileName,
-          shortPath,
-          data: parsedData,
-          isAddFile: true,
-        });
-        return;
-      }
-      if (ext === '.pts') {
-        const ptsData = await vscode.workspace.fs.readFile(fileUri);
-        const ptsParser = new PtsParser();
-        const parsedData = await ptsParser.parse(ptsData);
-        webviewPanel.webview.postMessage({
-          type: 'ptsData',
-          fileName,
-          shortPath,
-          data: parsedData,
-          isAddFile: true,
-        });
-        return;
-      }
-      if (ext === '.off') {
-        const offData = await vscode.workspace.fs.readFile(fileUri);
-        const offParser = new OffParser();
-        const parsedData = await offParser.parse(offData);
-        webviewPanel.webview.postMessage({
-          type: 'offData',
-          fileName,
-          shortPath,
-          data: parsedData,
-          isAddFile: true,
-        });
-        return;
-      }
-      if (ext === '.gltf' || ext === '.glb') {
-        const gltfData = await vscode.workspace.fs.readFile(fileUri);
-        const gltfParser = new GltfParser();
-        const parsedData = await gltfParser.parse(gltfData);
-        webviewPanel.webview.postMessage({
-          type: 'gltfData',
-          fileName,
-          shortPath,
-          data: parsedData,
-          isAddFile: true,
-        });
-        return;
-      }
-      if (ext === '.xyzn' || ext === '.xyzrgb') {
-        const xyzData = await vscode.workspace.fs.readFile(fileUri);
-        webviewPanel.webview.postMessage({
-          type: 'xyzVariantData',
-          fileName,
-          shortPath,
-          data: xyzData.buffer.slice(xyzData.byteOffset, xyzData.byteOffset + xyzData.byteLength),
-          variant: ext.substring(1),
-          isAddFile: true,
-        });
-        return;
-      }
-    } catch (error) {
-      vscode.window.showErrorMessage(
-        `Failed to add file from path: ${error instanceof Error ? error.message : String(error)}`
-      );
-    }
-  }
-
-  private async handleDroppedFilesFromWebview(
-    webviewPanel: vscode.WebviewPanel,
-    files: Array<{ name?: string; data?: ArrayBuffer | Uint8Array }>
-  ): Promise<void> {
-    for (let i = 0; i < files.length; i++) {
-      const droppedFile = files[i];
-      const fileName = path.basename(droppedFile.name || `dropped-file-${i + 1}`);
-
-      try {
-        if (!droppedFile.data) {
-          throw new Error('Dropped file did not include readable data');
-        }
-
-        const fileData = this.toUint8Array(droppedFile.data);
-        const shortPath = fileName;
-        const ext = path.extname(fileName).toLowerCase();
-        const fileType = detectFileTypeWithContent(fileName, fileData);
-
-        if (!fileType) {
-          webviewPanel.webview.postMessage({
-            type: 'loadingError',
-            fileName,
-            fileType: ext || 'file',
-            error: `Unsupported dropped file type: ${ext || 'unknown'}`,
-          });
-          continue;
-        }
-
-        if (fileType.extension === 'npy' && fileType.category === 'pointCloud') {
-          const npyParser = new NpyParser();
-          const parsedData = await npyParser.parse(fileData);
-          webviewPanel.webview.postMessage({
-            type: 'npyData',
-            fileName,
-            shortPath,
-            data: parsedData,
-            isAddFile: true,
-          });
-          continue;
-        }
-
-        if (fileType.isDepthFile) {
-          webviewPanel.webview.postMessage({
-            type: 'depthData',
-            fileName,
-            shortPath,
-            data: this.toArrayBuffer(fileData),
-            isAddFile: true,
-          });
-          continue;
-        }
-
-        if (ext === '.ply') {
-          const parser = new PlyParser();
-          const isBinary = isPlyBinary(fileData);
-          if (isBinary) {
-            const headerResult = await parser.parseHeaderOnly(fileData);
-            headerResult.headerInfo.fileName = fileName;
-            headerResult.headerInfo.shortPath = shortPath;
-            await this.sendUltimateRawBinary(
-              webviewPanel,
-              headerResult.headerInfo,
-              headerResult,
-              fileData,
-              'addFiles'
-            );
-          } else {
-            const parsedData = await parser.parse(fileData);
-            parsedData.fileName = fileName;
-            parsedData.shortPath = shortPath;
-            await this.sendSpatialDataToWebview(webviewPanel, [parsedData], 'addFiles');
-          }
-          continue;
-        }
-
-        if (ext === '.xyz') {
-          webviewPanel.webview.postMessage({
-            type: 'xyzData',
-            fileName,
-            shortPath,
-            data: this.toArrayBuffer(fileData),
-            isAddFile: true,
-          });
-          continue;
-        }
-
-        if (ext === '.obj') {
-          const objParser = new ObjParser();
-          const parsedData = await objParser.parse(fileData);
-          webviewPanel.webview.postMessage({
-            type: 'objData',
-            fileName,
-            shortPath,
-            data: parsedData,
-            isAddFile: true,
-          });
-          continue;
-        }
-
-        if (ext === '.stl') {
-          const stlParser = new StlParser();
-          const parsedData = await stlParser.parse(fileData);
-          webviewPanel.webview.postMessage({
-            type: 'stlData',
-            fileName,
-            shortPath,
-            data: parsedData,
-            isAddFile: true,
-          });
-          continue;
-        }
-
-        if (ext === '.pcd') {
-          const pcdParser = new PcdParser();
-          const parsedData = await pcdParser.parse(fileData);
-          webviewPanel.webview.postMessage({
-            type: 'pcdData',
-            fileName,
-            shortPath,
-            data: parsedData,
-            isAddFile: true,
-          });
-          continue;
-        }
-
-        if (ext === '.pts') {
-          const ptsParser = new PtsParser();
-          const parsedData = await ptsParser.parse(fileData);
-          webviewPanel.webview.postMessage({
-            type: 'ptsData',
-            fileName,
-            shortPath,
-            data: parsedData,
-            isAddFile: true,
-          });
-          continue;
-        }
-
-        if (ext === '.off') {
-          const offParser = new OffParser();
-          const parsedData = await offParser.parse(fileData);
-          webviewPanel.webview.postMessage({
-            type: 'offData',
-            fileName,
-            shortPath,
-            data: parsedData,
-            isAddFile: true,
-          });
-          continue;
-        }
-
-        if (ext === '.gltf' || ext === '.glb') {
-          const gltfParser = new GltfParser();
-          const parsedData = await gltfParser.parse(fileData);
-          webviewPanel.webview.postMessage({
-            type: 'gltfData',
-            fileName,
-            shortPath,
-            data: parsedData,
-            isAddFile: true,
-          });
-          continue;
-        }
-
-        if (ext === '.xyzn' || ext === '.xyzrgb') {
-          webviewPanel.webview.postMessage({
-            type: 'xyzVariantData',
-            fileName,
-            shortPath,
-            data: this.toArrayBuffer(fileData),
-            variant: ext.substring(1),
-            isAddFile: true,
-          });
-          continue;
-        }
-      } catch (error) {
-        webviewPanel.webview.postMessage({
-          type: 'loadingError',
-          fileName,
-          fileType: path.extname(fileName).slice(1).toUpperCase() || 'file',
-          error: error instanceof Error ? error.message : String(error),
-        });
-      }
-    }
-  }
-
-  private toUint8Array(data: ArrayBuffer | Uint8Array): Uint8Array {
-    if (data instanceof Uint8Array) {
-      return data;
-    }
-
-    return new Uint8Array(data);
-  }
-
-  private toArrayBuffer(data: Uint8Array): ArrayBuffer {
-    return data.buffer.slice(data.byteOffset, data.byteOffset + data.byteLength) as ArrayBuffer;
-  }
-
-  /**
-   * Read a file's bytes. For local files this uses Node's fs directly, which is
-   * markedly faster than vscode.workspace.fs on large files (the latter routes
-   * through the FS-provider abstraction). Falls back to workspace.fs for
-   * non-local/virtual schemes.
-   */
-  private async readFileFast(uri: vscode.Uri): Promise<Uint8Array> {
-    if (uri.scheme === 'file') {
-      try {
-        return await fs.promises.readFile(uri.fsPath);
-      } catch {
-        /* fall back to the workspace filesystem */
-      }
-    }
-    return vscode.workspace.fs.readFile(uri);
-  }
-
-  /**
-   * Read just the first `maxBytes` of a local file (for header detection before
-   * deciding whether to stream the body). Returns a copy sized to bytes actually
-   * read; throws on non-file schemes / IO errors (callers fall back).
-   */
-  private async readFileHead(uri: vscode.Uri, maxBytes: number): Promise<Uint8Array> {
-    const fh = await fs.promises.open(uri.fsPath, 'r');
-    try {
-      const buf = Buffer.allocUnsafe(maxBytes);
-      const { bytesRead } = await fh.read(buf, 0, maxBytes, 0);
-      return new Uint8Array(buf.subarray(0, bytesRead));
-    } finally {
-      await fh.close();
-    }
-  }
-
-  /**
-   * True if a PCD's VIEWPOINT is identity (or absent) — i.e. no rigid transform
-   * needs applying. The WASM PCD parser doesn't carry the viewpoint, so clouds
-   * with a non-identity viewpoint stay on the JS path that applies it.
-   */
-  private pcdViewpointIsIdentity(bytes: Uint8Array): boolean {
-    const head = Buffer.from(bytes.subarray(0, Math.min(1024, bytes.length))).toString('utf8');
-    const m = head.match(/^VIEWPOINT\s+(.+)$/m);
-    if (!m) {
-      return true;
-    }
-    const v = m[1].trim().split(/\s+/).map(Number);
-    return (
-      v.length >= 7 &&
-      v[0] === 0 &&
-      v[1] === 0 &&
-      v[2] === 0 &&
-      v[3] === 1 &&
-      v[4] === 0 &&
-      v[5] === 0 &&
-      v[6] === 0
-    );
   }
 
   private async handleSequenceRequestFile(
@@ -2154,816 +672,16 @@ export class PointCloudEditorProvider implements vscode.CustomReadonlyEditorProv
       parsedData.fileName = path.basename(uri.fsPath);
       parsedData.shortPath = this.getShortPath(uri.fsPath);
       parsedData.fileIndex = 0;
-      await this.sendUltimateRawBinary(
+      await sendUltimateRawBinary(
         panel,
         parsedData,
         headerResult,
         bytes,
-        message.messageType || 'multiSpatialData'
+        message.messageType || 'multiSpatialData',
+        this.logPerf.bind(this)
       );
     } catch (error) {
       console.error('PLY fetch fallback failed:', error);
-    }
-  }
-
-  private async sendUltimateRawBinary(
-    webviewPanel: vscode.WebviewPanel,
-    parsedData: any,
-    headerResult: any,
-    rawFileData: Uint8Array,
-    messageType: string
-  ): Promise<void> {
-    console.log(`🚀 ULTIMATE: Sending raw binary data for ${parsedData.fileName}`);
-
-    // Extract exactly the binary vertex bytes into a standalone ArrayBuffer.
-    // NOTE: vscode.workspace.fs.readFile returns a Node Buffer, whose .slice()
-    // is a *view* over the whole file's ArrayBuffer — so we must slice .buffer
-    // by byteOffset/byteLength to copy out just the vertex region. Sending the
-    // raw .buffer instead would ship the whole file from offset 0 (header bytes
-    // read as float32 → garbage geometry).
-    const copyStart = performance.now();
-    const binaryVertexData = rawFileData.subarray(headerResult.binaryDataStart);
-    const slicedBuffer = binaryVertexData.buffer.slice(
-      binaryVertexData.byteOffset,
-      binaryVertexData.byteOffset + binaryVertexData.byteLength
-    );
-    const copyMs = performance.now() - copyStart;
-    this.logPerf(
-      `⏱️ PERF[ply/ext] copy ${copyMs.toFixed(1)}ms (${(binaryVertexData.byteLength / 1048576).toFixed(1)}MB) for ${parsedData.fileName}`
-    );
-
-    // Send raw binary data + parsing metadata
-    webviewPanel.webview.postMessage({
-      type: 'ultimateRawBinaryData',
-      messageType: messageType,
-      // Wall-clock epoch stamped right before postMessage so the webview can
-      // measure the cross-process serialization+IPC "transfer" cost.
-      postedAt: Date.now(),
-      fileName: parsedData.fileName,
-      shortPath: parsedData.shortPath,
-      fileSizeInBytes: rawFileData.byteLength,
-      vertexCount: parsedData.vertexCount,
-      faceCount: parsedData.faceCount,
-      hasColors: parsedData.hasColors,
-      hasNormals: parsedData.hasNormals,
-      hasIntensity: parsedData.hasIntensity,
-      format: parsedData.format,
-      comments: parsedData.comments,
-
-      // Raw binary data + parsing info
-      rawBinaryData: slicedBuffer,
-      vertexStride: headerResult.vertexStride,
-      propertyOffsets: Array.from(headerResult.propertyOffsets.entries()),
-      littleEndian: headerResult.headerInfo.format === 'binary_little_endian',
-      faceCountType: headerResult.faceCountType,
-      faceIndexType: headerResult.faceIndexType,
-    });
-  }
-
-  private async sendSpatialDataToWebview(
-    webviewPanel: vscode.WebviewPanel,
-    spatialDataArray: any[],
-    messageType: string
-  ): Promise<void> {
-    for (const spatialData of spatialDataArray) {
-      console.log(
-        `🚀 Binary transfer for ${spatialData.fileName} (${spatialData.vertexCount} vertices)`
-      );
-      const startTime = performance.now();
-
-      try {
-        await this.sendBinaryData(webviewPanel, spatialData, messageType);
-        const transferTime = performance.now() - startTime;
-        console.log(`⚡ Binary transfer complete: ${transferTime.toFixed(1)}ms`);
-      } catch (error) {
-        console.log(
-          `⚠️ Binary transfer failed for ${spatialData.fileName}, falling back to chunking...`
-        );
-        await this.sendLargeFileInChunksOptimized(webviewPanel, spatialData, messageType);
-      }
-    }
-  }
-
-  private async sendBinaryData(
-    webviewPanel: vscode.WebviewPanel,
-    spatialData: any,
-    messageType: string
-  ): Promise<void> {
-    // Check if we have direct TypedArrays (ultra-fast path)
-    const vertexCount = spatialData.vertexCount;
-    const hasColors = spatialData.hasColors;
-    const hasNormals = spatialData.hasNormals;
-
-    let positionBuffer: Float32Array;
-    let colorBuffer: Uint8Array | null = null;
-    let normalBuffer: Float32Array | null = null;
-
-    if (spatialData.useTypedArrays) {
-      // Ultra-fast: Use TypedArrays directly (zero-copy)
-      console.log(`🚀 Using direct TypedArrays for binary transfer - ZERO COPY!`);
-      positionBuffer = spatialData.positionsArray;
-      colorBuffer = spatialData.colorsArray;
-      normalBuffer = spatialData.normalsArray;
-    } else {
-      // Fallback: Convert vertex objects to TypedArrays
-      console.log(`🔄 Converting vertex objects to TypedArrays for binary transfer...`);
-      const vertices = spatialData.vertices;
-
-      // Create typed arrays for vertices (always 3 floats: x, y, z)
-      positionBuffer = new Float32Array(vertexCount * 3);
-
-      // Optional color buffer (RGB as bytes: 0-255)
-      if (hasColors) {
-        colorBuffer = new Uint8Array(vertexCount * 3);
-      }
-
-      // Optional normal buffer (3 floats: nx, ny, nz)
-      if (hasNormals) {
-        normalBuffer = new Float32Array(vertexCount * 3);
-      }
-
-      // Fill the buffers
-      for (let i = 0; i < vertexCount; i++) {
-        const vertex = vertices[i];
-
-        // Position (always present)
-        positionBuffer[i * 3] = vertex.x;
-        positionBuffer[i * 3 + 1] = vertex.y;
-        positionBuffer[i * 3 + 2] = vertex.z;
-
-        // Colors (if present)
-        if (hasColors && colorBuffer) {
-          colorBuffer[i * 3] = vertex.red || 0;
-          colorBuffer[i * 3 + 1] = vertex.green || 0;
-          colorBuffer[i * 3 + 2] = vertex.blue || 0;
-        }
-
-        // Normals (if present)
-        if (hasNormals && normalBuffer) {
-          normalBuffer[i * 3] = vertex.nx || 0;
-          normalBuffer[i * 3 + 1] = vertex.ny || 0;
-          normalBuffer[i * 3 + 2] = vertex.nz || 0;
-        }
-      }
-    }
-
-    // Handle faces if present
-    let indexBuffer: Uint32Array | null = null;
-    if (spatialData.faces && spatialData.faces.length > 0) {
-      const faces = spatialData.faces;
-      indexBuffer = new Uint32Array(faces.length * 3); // Assuming triangles
-
-      for (let i = 0; i < faces.length; i++) {
-        const face = faces[i];
-        indexBuffer[i * 3] = face.indices[0];
-        indexBuffer[i * 3 + 1] = face.indices[1];
-        indexBuffer[i * 3 + 2] = face.indices[2];
-      }
-    }
-
-    // Calculate total binary size
-    const totalSize =
-      positionBuffer.byteLength +
-      (colorBuffer ? colorBuffer.byteLength : 0) +
-      (normalBuffer ? normalBuffer.byteLength : 0) +
-      (indexBuffer ? indexBuffer.byteLength : 0);
-
-    console.log(
-      `📦 Binary data: ${(totalSize / 1024 / 1024).toFixed(1)}MB (${vertexCount} vertices)`
-    );
-
-    // Send metadata + binary buffers
-    webviewPanel.webview.postMessage({
-      type: 'binarySpatialData',
-      messageType: messageType,
-      fileName: spatialData.fileName,
-      shortPath: spatialData.shortPath,
-      vertexCount: vertexCount,
-      faceCount: spatialData.faceCount,
-      hasColors: hasColors,
-      hasNormals: hasNormals,
-      format: spatialData.format,
-      comments: spatialData.comments,
-      // Binary buffers (will be transferred efficiently)
-      positionBuffer: positionBuffer.buffer,
-      colorBuffer: colorBuffer ? colorBuffer.buffer : null,
-      normalBuffer: normalBuffer ? normalBuffer.buffer : null,
-      indexBuffer: indexBuffer ? indexBuffer.buffer : null,
-    });
-  }
-
-  private async sendLargeFileInChunksOptimized(
-    webviewPanel: vscode.WebviewPanel,
-    spatialData: any,
-    messageType: string
-  ): Promise<void> {
-    // ULTRA-AGGRESSIVE chunking for maximum transfer speed
-    const CHUNK_SIZE = 1000000; // 1M vertices per chunk!
-    const totalVertices = spatialData.vertexCount;
-    const vertices = spatialData.vertices;
-    const colors = spatialData.colors;
-    const normals = spatialData.normals;
-    const faces = spatialData.faces;
-
-    const totalChunks = Math.ceil(totalVertices / CHUNK_SIZE);
-    console.log(
-      `🚀 Ultra-fast chunking: ${spatialData.fileName} (${totalVertices} vertices, ${totalChunks} chunks)`
-    );
-
-    const startTime = performance.now();
-    let firstChunkTime = 0;
-
-    // Send start message
-    webviewPanel.webview.postMessage({
-      type: 'startLargeFile',
-      fileName: spatialData.fileName,
-      shortPath: spatialData.shortPath,
-      totalVertices: totalVertices,
-      totalChunks: totalChunks,
-      hasColors: spatialData.hasColors,
-      hasNormals: spatialData.hasNormals,
-      faces: faces,
-      format: spatialData.format,
-      comments: spatialData.comments,
-      messageType: messageType,
-    });
-
-    // Send chunks with minimal overhead
-    for (let i = 0; i < totalChunks; i++) {
-      const startIdx = i * CHUNK_SIZE;
-      const endIdx = Math.min(startIdx + CHUNK_SIZE, totalVertices);
-      const chunkSize = endIdx - startIdx;
-
-      // Extract chunk data efficiently
-      const chunkVertices = vertices.slice(startIdx, endIdx);
-      const chunkColors = colors ? colors.slice(startIdx, endIdx) : undefined;
-      const chunkNormals = normals ? normals.slice(startIdx, endIdx) : undefined;
-
-      webviewPanel.webview.postMessage({
-        type: 'largeFileChunk',
-        fileName: spatialData.fileName,
-        chunkIndex: i,
-        totalChunks: totalChunks,
-        vertices: chunkVertices,
-        colors: chunkColors,
-        normals: chunkNormals,
-      });
-
-      if (i === 0) {
-        firstChunkTime = performance.now();
-      }
-
-      // Log only every 5th chunk to reduce console spam
-      if (i % 5 === 0 || i === totalChunks - 1) {
-        console.log(`Chunk ${i + 1}/${totalChunks} (${chunkSize} vertices)`);
-      }
-    }
-
-    // Send completion message
-    webviewPanel.webview.postMessage({
-      type: 'largeFileComplete',
-      fileName: spatialData.fileName,
-      messageType: messageType,
-    });
-
-    const totalTime = performance.now() - startTime;
-    console.log(
-      `⚡ Ultra-fast transfer complete: ${totalTime.toFixed(1)}ms total, ${firstChunkTime ? (firstChunkTime - startTime).toFixed(1) : 0}ms to first chunk`
-    );
-  }
-
-  private async handleCameraParametersRequest(
-    webviewPanel: vscode.WebviewPanel,
-    message: any
-  ): Promise<void> {
-    try {
-      // Load saved default settings (filter out cx/cy as they should be auto-calculated per image)
-      const savedSettings = this.context.globalState.get('defaultDepthSettings') as any;
-      const defaults = savedSettings
-        ? {
-            fx: savedSettings.fx || 1000,
-            fy: savedSettings.fy,
-            cameraModel: savedSettings.cameraModel || 'pinhole-ideal',
-            depthType: savedSettings.depthType || 'euclidean',
-            convention: savedSettings.convention || 'opengl',
-            baseline: savedSettings.baseline || 50,
-            pngScaleFactor: savedSettings.pngScaleFactor || 1000,
-            // Explicitly exclude cx and cy
-          }
-        : {
-            fx: 1000,
-            fy: undefined,
-            cameraModel: 'pinhole-ideal',
-            depthType: 'euclidean',
-            convention: 'opengl',
-            baseline: 50,
-            pngScaleFactor: 1000,
-          };
-
-      console.log('🎯 Using default settings for camera parameters dialog:', defaults);
-
-      // Automatically use default settings without showing dialog
-      // (Original dialog code commented out - user requested to always use defaults)
-      /*
-      const useDefaults = await vscode.window.showQuickPick(
-        [
-          {
-            label: '⚡ Use Default Settings',
-            description: `${defaults.cameraModel}, fx=${defaults.fx}px${defaults.fy ? `, fy=${defaults.fy}px` : ''}, ${defaults.depthType}, ${defaults.convention}${defaults.baseline ? `, baseline=${defaults.baseline}mm` : ''}`,
-            value: 'defaults',
-          },
-          {
-            label: '⚙️ Customize Settings',
-            description: 'Choose settings manually',
-            value: 'customize',
-          },
-        ],
-        {
-          placeHolder: 'Convert depth image to point cloud',
-          ignoreFocusOut: true,
-        }
-      );
-
-      if (!useDefaults) {
-        webviewPanel.webview.postMessage({
-          type: 'cameraParamsCancelled',
-          requestId: message.requestId,
-        });
-        return;
-      }
-
-      if (useDefaults.value === 'defaults') {
-      */
-
-      // Always use defaults automatically
-      // Use saved defaults without showing additional dialogs
-      webviewPanel.webview.postMessage({
-        type: 'cameraParams',
-        cameraModel: defaults.cameraModel,
-        fx: defaults.fx,
-        fy: defaults.fy,
-        depthType: defaults.depthType,
-        baseline: defaults.baseline,
-        convention: defaults.convention,
-        requestId: message.requestId,
-      });
-      return;
-
-      /*
-      // All customization dialog code commented out (user requested to always use defaults)
-
-      // Show camera model selection dialog
-      const cameraModel = await vscode.window.showQuickPick(
-        [
-          {
-            label: 'Pinhole Camera',
-            description:
-              defaults.cameraModel === 'pinhole-ideal'
-                ? 'Standard perspective projection model (Default)'
-                : 'Standard perspective projection model',
-            value: 'pinhole-ideal',
-          },
-          {
-            label: 'Fisheye Camera',
-            description:
-              defaults.cameraModel === 'fisheye'
-                ? 'Wide-angle fisheye projection model (Default)'
-                : 'Wide-angle fisheye projection model',
-            value: 'fisheye',
-          },
-        ],
-        {
-          placeHolder: `Select camera model used to capture the depth image (Default: ${defaults.cameraModel})`,
-          ignoreFocusOut: true,
-        }
-      );
-
-      if (!cameraModel) {
-        webviewPanel.webview.postMessage({
-          type: 'cameraParamsCancelled',
-          requestId: message.requestId,
-        });
-        return;
-      }
-
-      // Show depth type selection dialog
-      const depthType = await vscode.window.showQuickPick(
-        [
-          {
-            label: 'Euclidean Depth',
-            description:
-              defaults.depthType === 'euclidean'
-                ? 'Metric depth values (distance from camera center) (Default)'
-                : 'Metric depth values (distance from camera center)',
-            value: 'euclidean',
-          },
-          {
-            label: 'Orthogonal Depth',
-            description:
-              defaults.depthType === 'orthogonal'
-                ? 'Z-buffer depth values (Z-coordinate) (Default)'
-                : 'Z-buffer depth values (Z-coordinate)',
-            value: 'orthogonal',
-          },
-          {
-            label: 'Disparity',
-            description:
-              defaults.depthType === 'disparity'
-                ? 'Disparity values (requires baseline parameter) (Default)'
-                : 'Disparity values (requires baseline parameter)',
-            value: 'disparity',
-          },
-        ],
-        {
-          placeHolder: `Select the type of depth data in your image (Default: ${defaults.depthType})`,
-          ignoreFocusOut: true,
-        }
-      );
-
-      if (!depthType) {
-        webviewPanel.webview.postMessage({
-          type: 'cameraParamsCancelled',
-          requestId: message.requestId,
-        });
-        return;
-      }
-
-      // Show focal length input dialog
-      const fxInput = await vscode.window.showInputBox({
-        prompt: `Enter fx (focal length x) in pixels (Default: ${defaults.fx})`,
-        placeHolder: defaults.fx.toString(),
-        value: defaults.fx.toString(),
-        validateInput: (value: string) => {
-          const num = parseFloat(value);
-          if (isNaN(num) || num <= 0) {
-            return 'Please enter a valid positive number for fx';
-          }
-          return null;
-        },
-        ignoreFocusOut: true,
-      });
-
-      if (!fxInput) {
-        webviewPanel.webview.postMessage({
-          type: 'cameraParamsCancelled',
-          requestId: message.requestId,
-        });
-        return;
-      }
-
-      const fx = parseFloat(fxInput);
-
-      // Show fy input dialog (optional)
-      const fyInput = await vscode.window.showInputBox({
-        prompt: `Enter fy (focal length y) in pixels (Default: same as fx = ${fx})`,
-        placeHolder: 'Leave empty to use same as fx',
-        value: defaults.fy?.toString() || '',
-        validateInput: (value: string) => {
-          if (value.trim() === '') {
-            return null;
-          } // Empty is OK
-          const num = parseFloat(value);
-          if (isNaN(num) || num <= 0) {
-            return 'Please enter a valid positive number for fy, or leave empty';
-          }
-          return null;
-        },
-        ignoreFocusOut: true,
-      });
-
-      if (fyInput === undefined) {
-        // User cancelled
-        webviewPanel.webview.postMessage({
-          type: 'cameraParamsCancelled',
-          requestId: message.requestId,
-        });
-        return;
-      }
-
-      const fy = fyInput.trim() === '' ? undefined : parseFloat(fyInput);
-
-      // Show coordinate convention selection dialog
-      const convention = await vscode.window.showQuickPick(
-        [
-          {
-            label: 'OpenGL Convention (Y-up, Z-backward)',
-            description:
-              defaults.convention === 'opengl'
-                ? 'Standard 3D graphics convention (Default)'
-                : 'Standard 3D graphics convention',
-            value: 'opengl',
-          },
-          {
-            label: 'OpenCV Convention (Y-down, Z-forward)',
-            description:
-              defaults.convention === 'opencv'
-                ? 'Computer vision convention (Default)'
-                : 'Computer vision convention',
-            value: 'opencv',
-          },
-        ],
-        {
-          placeHolder: `Select coordinate convention for the resulting point cloud (Default: ${defaults.convention})`,
-          ignoreFocusOut: true,
-        }
-      );
-
-      if (!convention) {
-        webviewPanel.webview.postMessage({
-          type: 'cameraParamsCancelled',
-          requestId: message.requestId,
-        });
-        return;
-      }
-
-      // Show baseline input dialog if disparity is selected
-      let baseline: number | undefined;
-      if (depthType.value === 'disparity') {
-        const defaultBaseline = defaults.baseline || 50;
-        const baselineInput = await vscode.window.showInputBox({
-          prompt: `Enter the baseline in millimeters (Default: ${defaultBaseline})`,
-          placeHolder: defaultBaseline.toString(),
-          value: defaultBaseline.toString(),
-          validateInput: (value: string) => {
-            const num = parseFloat(value);
-            if (isNaN(num) || num <= 0) {
-              return 'Please enter a valid positive number for baseline';
-            }
-            return null;
-          },
-          ignoreFocusOut: true,
-        });
-
-        if (!baselineInput) {
-          webviewPanel.webview.postMessage({
-            type: 'cameraParamsCancelled',
-            requestId: message.requestId,
-          });
-          return;
-        }
-
-        baseline = parseFloat(baselineInput);
-      }
-
-      // Send camera parameters to webview
-      webviewPanel.webview.postMessage({
-        type: 'cameraParams',
-        cameraModel: cameraModel.value,
-        fx: fx,
-        fy: fy,
-        depthType: depthType.value,
-        baseline: baseline,
-        convention: convention.value,
-        requestId: message.requestId,
-      });
-      */
-    } catch (error) {
-      webviewPanel.webview.postMessage({
-        type: 'cameraParamsError',
-        error: error instanceof Error ? error.message : String(error),
-        requestId: message.requestId,
-      });
-    }
-  }
-
-  private async handleCameraParametersWithScaleRequest(
-    webviewPanel: vscode.WebviewPanel,
-    message: any
-  ): Promise<void> {
-    try {
-      // Load saved default settings (filter out cx/cy as they should be auto-calculated per image)
-      const savedSettings = this.context.globalState.get('defaultDepthSettings') as any;
-      const defaults = savedSettings
-        ? {
-            fx: savedSettings.fx || 1000,
-            fy: savedSettings.fy,
-            cameraModel: savedSettings.cameraModel || 'pinhole-ideal',
-            depthType: savedSettings.depthType || 'euclidean',
-            convention: savedSettings.convention || 'opengl',
-            baseline: savedSettings.baseline || 50,
-            pngScaleFactor: savedSettings.pngScaleFactor || 1000,
-            // Explicitly exclude cx and cy
-          }
-        : {
-            fx: 1000,
-            fy: undefined,
-            cameraModel: 'pinhole-ideal',
-            depthType: 'euclidean',
-            convention: 'opengl',
-            baseline: 50,
-            pngScaleFactor: 1000, // Default for PNG: millimeters to meters
-          };
-
-      console.log('🎯 Using default settings for PNG camera parameters dialog:', defaults);
-
-      // Show option to use defaults directly or customize
-      const useDefaults = await vscode.window.showQuickPick(
-        [
-          {
-            label: '⚡ Use Default Settings',
-            description: `${defaults.cameraModel}, fx=${defaults.fx}px${defaults.fy ? `, fy=${defaults.fy}px` : ''}, scale=${defaults.pngScaleFactor} (${defaults.pngScaleFactor === 1000 ? 'mm→m' : defaults.pngScaleFactor === 256 ? 'disp÷256' : 'custom'})`,
-            value: 'defaults',
-          },
-          {
-            label: '⚙️ Customize Settings',
-            description: 'Choose settings manually',
-            value: 'customize',
-          },
-        ],
-        {
-          placeHolder: 'Convert PNG depth image to point cloud',
-          ignoreFocusOut: true,
-        }
-      );
-
-      if (!useDefaults) {
-        webviewPanel.webview.postMessage({
-          type: 'cameraParamsCancelled',
-          requestId: message.requestId,
-        });
-        return;
-      }
-
-      if (useDefaults.value === 'defaults') {
-        // Use saved defaults without showing additional dialogs
-        webviewPanel.webview.postMessage({
-          type: 'cameraParams',
-          cameraModel: defaults.cameraModel,
-          fx: defaults.fx,
-          fy: defaults.fy,
-          depthType: defaults.depthType,
-          baseline: defaults.baseline,
-          convention: defaults.convention,
-          pngScaleFactor: defaults.pngScaleFactor,
-          requestId: message.requestId,
-        });
-        return;
-      }
-
-      // Show camera model selection dialog
-      const cameraModel = await vscode.window.showQuickPick(
-        [
-          {
-            label: 'Pinhole Camera',
-            description:
-              defaults.cameraModel === 'pinhole-ideal'
-                ? 'Standard perspective projection model (Default)'
-                : 'Standard perspective projection model',
-            value: 'pinhole-ideal',
-          },
-          {
-            label: 'Fisheye Camera',
-            description:
-              defaults.cameraModel === 'fisheye'
-                ? 'Wide-angle fisheye projection model (Default)'
-                : 'Wide-angle fisheye projection model',
-            value: 'fisheye',
-          },
-        ],
-        {
-          placeHolder: `Select camera model used to capture the depth image (Default: ${defaults.cameraModel})`,
-          ignoreFocusOut: true,
-        }
-      );
-
-      if (!cameraModel) {
-        webviewPanel.webview.postMessage({
-          type: 'cameraParamsCancelled',
-          requestId: message.requestId,
-        });
-        return;
-      }
-
-      // Show scale factor input dialog
-      const pngScaleFactorInput = await vscode.window.showInputBox({
-        prompt: `Scale factor: depth/disparity is divided to get applied value in meters/disparities (Default: ${defaults.pngScaleFactor})`,
-        placeHolder: `${defaults.pngScaleFactor} (1000 for mm, 256 for disparity, 1 for meters)`,
-        value: defaults.pngScaleFactor.toString(),
-        validateInput: (value: string) => {
-          const num = parseFloat(value);
-          if (isNaN(num) || num <= 0) {
-            return 'Please enter a valid positive number for scale factor';
-          }
-          return null;
-        },
-        ignoreFocusOut: true,
-      });
-
-      if (!pngScaleFactorInput) {
-        webviewPanel.webview.postMessage({
-          type: 'cameraParamsCancelled',
-          requestId: message.requestId,
-        });
-        return;
-      }
-
-      const pngScaleFactor = parseFloat(pngScaleFactorInput);
-
-      // Show focal length input dialog
-      const fxInput = await vscode.window.showInputBox({
-        prompt: `Enter fx (focal length x) in pixels (Default: ${defaults.fx})`,
-        placeHolder: defaults.fx.toString(),
-        value: defaults.fx.toString(),
-        validateInput: (value: string) => {
-          const num = parseFloat(value);
-          if (isNaN(num) || num <= 0) {
-            return 'Please enter a valid positive number for fx';
-          }
-          return null;
-        },
-        ignoreFocusOut: true,
-      });
-
-      if (!fxInput) {
-        webviewPanel.webview.postMessage({
-          type: 'cameraParamsCancelled',
-          requestId: message.requestId,
-        });
-        return;
-      }
-
-      const fx = parseFloat(fxInput);
-
-      // Show fy input dialog (optional)
-      const fyInput = await vscode.window.showInputBox({
-        prompt: `Enter fy (focal length y) in pixels (Default: same as fx = ${fx})`,
-        placeHolder: 'Leave empty to use same as fx',
-        value: defaults.fy?.toString() || '',
-        validateInput: (value: string) => {
-          if (value.trim() === '') {
-            return null;
-          } // Empty is OK
-          const num = parseFloat(value);
-          if (isNaN(num) || num <= 0) {
-            return 'Please enter a valid positive number for fy, or leave empty';
-          }
-          return null;
-        },
-        ignoreFocusOut: true,
-      });
-
-      if (fyInput === undefined) {
-        // User cancelled
-        webviewPanel.webview.postMessage({
-          type: 'cameraParamsCancelled',
-          requestId: message.requestId,
-        });
-        return;
-      }
-
-      const fy = fyInput.trim() === '' ? undefined : parseFloat(fyInput);
-
-      // Show coordinate convention selection dialog
-      const convention = await vscode.window.showQuickPick(
-        [
-          {
-            label: 'OpenGL Convention (Y-up, Z-backward)',
-            description:
-              defaults.convention === 'opengl'
-                ? 'Standard 3D graphics convention (Default)'
-                : 'Standard 3D graphics convention',
-            value: 'opengl',
-          },
-          {
-            label: 'OpenCV Convention (Y-down, Z-forward)',
-            description:
-              defaults.convention === 'opencv'
-                ? 'Computer vision convention (Default)'
-                : 'Computer vision convention',
-            value: 'opencv',
-          },
-        ],
-        {
-          placeHolder: `Select coordinate convention for the resulting point cloud (Default: ${defaults.convention})`,
-          ignoreFocusOut: true,
-        }
-      );
-
-      if (!convention) {
-        webviewPanel.webview.postMessage({
-          type: 'cameraParamsCancelled',
-          requestId: message.requestId,
-        });
-        return;
-      }
-
-      // Send camera parameters to webview
-      webviewPanel.webview.postMessage({
-        type: 'cameraParams',
-        cameraModel: cameraModel.value,
-        fx: fx,
-        fy: fy,
-        depthType: 'euclidean', // Default for PNG
-        pngScaleFactor: pngScaleFactor,
-        convention: convention.value,
-        requestId: message.requestId,
-      });
-    } catch (error) {
-      webviewPanel.webview.postMessage({
-        type: 'cameraParamsError',
-        error: error instanceof Error ? error.message : String(error),
-        requestId: message.requestId,
-      });
     }
   }
